Add headerView to TableView
