+++ Bankey/AccountSummary/AccountSummaryHeaderView.xib
@@ -17,15 +17,65 @@
         <view contentMode="scaleToFill" id="iN0-l3-epB">
             <rect key="frame" x="0.0" y="0.0" width="370" height="144"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Bankey" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="bmx-Dx-qEL">
+                    <rect key="frame" x="16" y="16" width="89" height="34"/>
+                    <fontDescription key="fontDescription" style="UICTFontTextStyleTitle1"/>
+                    <nil key="textColor"/>
+                    <nil key="highlightedColor"/>
+                </label>
+                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Date" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="XAQ-fq-TNM">
+                    <rect key="frame" x="16" y="116" width="36" height="21"/>
+                    <fontDescription key="fontDescription" style="UICTFontTextStyleBody"/>
+                    <nil key="textColor"/>
+                    <nil key="highlightedColor"/>
+                </label>
+                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Good morning" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="CP4-6s-pLj">
+                    <rect key="frame" x="16" y="58" width="109" height="21"/>
+                    <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                    <nil key="textColor"/>
+                    <nil key="highlightedColor"/>
+                </label>
+                <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="sun.max.fill" catalog="system" translatesAutoresizingMaskIntoConstraints="NO" id="g9b-IL-PNO">
+                    <rect key="frame" x="254" y="16" width="100" height="100"/>
+                    <color key="tintColor" systemColor="systemYellowColor"/>
+                    <constraints>
+                        <constraint firstAttribute="height" constant="100" id="dhj-pD-6gc"/>
+                        <constraint firstAttribute="width" constant="100" id="obe-TP-Iin"/>
+                    </constraints>
+                </imageView>
+                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Kevin" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="xLd-xI-V9c">
+                    <rect key="frame" x="16" y="87" width="42" height="21"/>
+                    <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                    <nil key="textColor"/>
+                    <nil key="highlightedColor"/>
+                </label>
+            </subviews>
             <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
             <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+            <constraints>
+                <constraint firstItem="CP4-6s-pLj" firstAttribute="top" secondItem="bmx-Dx-qEL" secondAttribute="bottom" constant="8" id="B3v-78-GNy"/>
+                <constraint firstItem="XAQ-fq-TNM" firstAttribute="top" secondItem="xLd-xI-V9c" secondAttribute="bottom" constant="8" id="DDr-yo-50g"/>
+                <constraint firstItem="bmx-Dx-qEL" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="16" id="ET0-TL-FOo"/>
+                <constraint firstItem="g9b-IL-PNO" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="16" id="JXy-hX-rmU"/>
+                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="g9b-IL-PNO" secondAttribute="trailing" constant="16" id="LfQ-FT-6yM"/>
+                <constraint firstItem="bmx-Dx-qEL" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="16" id="Qnt-ai-yM2"/>
+                <constraint firstItem="xLd-xI-V9c" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="16" id="UAP-Dq-BYd"/>
+                <constraint firstItem="xLd-xI-V9c" firstAttribute="top" secondItem="CP4-6s-pLj" secondAttribute="bottom" constant="8" id="bhR-eg-cYO"/>
+                <constraint firstItem="XAQ-fq-TNM" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="16" id="mU6-SG-rxt"/>
+                <constraint firstItem="CP4-6s-pLj" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="16" id="wip-rh-Fjc"/>
+            </constraints>
             <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <point key="canvasLocation" x="111" y="21"/>
+            <point key="canvasLocation" x="110.76923076923076" y="20.616113744075829"/>
         </view>
     </objects>
     <resources>
+        <image name="sun.max.fill" catalog="system" width="128" height="125"/>
         <systemColor name="systemBackgroundColor">
             <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+        </systemColor>
+        <systemColor name="systemYellowColor">
+            <color red="1" green="0.80000000000000004" blue="0.0" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
         </systemColor>
     </resources>
 </document>
