--- Bankey/AccountSummary/AccountSummaryViewController.swift
@@ -1,69 +1,0 @@
-//
-//  AccountSummaryViewController.swift
-//  Bankey
-//
-//  Created by User 2 on 06/10/22.
-//
-
-import UIKit
-
-class AccountSummaryViewController: UIViewController{
-    
-    let names = ["Sardor","Karshiev","Ayubjon O`g`li"]
-    var tableView = UITableView()
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        setup()
-    }
-}
-
-extension AccountSummaryViewController{
-    func setup(){
-        setupTableView()
-        setupTableHeaderView()
-    }
-    
-    func setupTableView(){
-        tableView.dataSource = self
-        tableView.delegate = self
-        
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(tableView)
-        
-        NSLayoutConstraint.activate([
-            tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-            tableView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
-            tableView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor),
-            tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-        ])
-    }
-    
-    private func setupTableHeaderView(){
-        let header = AccountSummaryHeaderView(frame: .zero)
-        
-        var size = header.systemLayoutSizeFitting(UIView.layoutFittingCompressedSize)
-        size.width = UIScreen.main.bounds.width
-        
-        tableView.tableHeaderView = header
-    }
-}
-
-extension AccountSummaryViewController: UITableViewDataSource,UITableViewDelegate{
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return names.count
-    }
-    
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-            let cell = UITableViewCell()
-            cell.textLabel?.text = names[indexPath.row]
-            return cell
-    }
-    
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        
-    }
-}
-
